sc build redesign - solution using separate wasm-output crate
