+++ Cargo.toml
@@ -4,5 +4,6 @@ members = [
   "elrond-wasm-debug",
   "elrond-wasm-derive",
   "elrond-wasm-node",
+  "elrond-wasm-output",
   "elrond-wasm-serde",
 ]
