+++ elrond-wasm-node/src/ext_error.rs
